requirements.txt
replicate
requests
spacy
python-dotenv